requirements.txt
bs4
googlevoice
pyyaml
requests
sqlalchemy
